ADD : Create road junction function (closes #1661)
git-svn-id: http://svn.openstreetmap.org/applications/editors/merkaartor@14861 b9d5c4c9-76e1-0310-9c85-f3177eceb1e4

+++ src/Maps/TrackPoint.h
@@ -91,7 +91,7 @@ Q_DECLARE_METATYPE( TrackPoint * );
 
 #ifndef _MOBILE
 // Register this point as being a recognizable point by the GGL
-GEOMETRY_REGISTER_POINT_2D_CONST(TrackPoint, int, cs::cartesian, projection().x(), projection().y())
+GEOMETRY_REGISTER_POINT_2D_CONST(TrackPoint, qreal, cs::cartesian, projection().x(), projection().y())
 #endif
